support #2: v2ray full support

Closes #2

See merge request self-code/api-server!8

+++ persist/persist.go
@@ -2,7 +2,10 @@ package persist
 
 import (
 	"context"
+	"fmt"
 	"time"
+
+	"cntechpower.com/api-server/log"
 
 	"cntechpower.com/api-server/model"
 	"github.com/go-redis/redis/v8"
@@ -13,7 +16,10 @@ import (
 var db *gorm.DB
 var cache *redis.Client
 
+var ErrDBNotInit = fmt.Errorf("mysql is not init")
+
 func Init(mysqlDsn, redisDsn string) error {
+	header := log.NewHeader("persist.Init")
 	var err error
 	db, err = gorm.Open(mysql.Open(mysqlDsn), &gorm.Config{})
 	if err != nil {
@@ -22,11 +28,18 @@ func Init(mysqlDsn, redisDsn string) error {
 	if err := db.AutoMigrate(model.GetAllModels()...); err != nil {
 		return err
 	}
-	cache = redis.NewClient(&redis.Options{
-		Addr:     redisDsn,
-		Password: "", // no password set
-		DB:       0,  // use default DB
-	})
+	log.Infof(header, "init mysql success")
+	if redisDsn != "" {
+		cache = redis.NewClient(&redis.Options{
+			Addr:     redisDsn,
+			Password: "", // no password set
+			DB:       0,  // use default DB
+		})
+		if err := cache.Ping(context.Background()).Err(); err != nil {
+			return err
+		}
+		log.Infof(header, "init redis success")
+	}
 	return nil
 }
 
@@ -39,55 +52,79 @@ func Redis() *redis.Client {
 }
 
 func Get(m model.Modeler) error {
-	if err := cache.Get(context.Background(), m.GetCacheKey()).Scan(m); err == nil {
-		return nil
+	if cache != nil {
+		if err := cache.Get(context.Background(), m.GetCacheKey()).Scan(m); err == nil {
+			return nil
+		}
 	}
 	if err := db.Find(m).Error; err != nil {
 		return err
 	}
-	cache.Set(context.Background(), m.GetCacheKey(), m, m.GetCacheDuration())
+	if cache != nil {
+		cache.Set(context.Background(), m.GetCacheKey(), m, m.GetCacheDuration())
+	}
 	return nil
 }
 
 func Create(m model.Modeler) error {
+	if db == nil {
+		return ErrDBNotInit
+	}
 	if err := db.Create(m).Error; err != nil {
 		return err
 	}
-	cache.Set(context.Background(), m.GetCacheKey(), m, m.GetCacheDuration())
+	if cache != nil {
+		cache.Set(context.Background(), m.GetCacheKey(), m, m.GetCacheDuration())
+	}
 	return nil
 }
 
 func Delete(m model.Modeler) error {
+	if db == nil {
+		return ErrDBNotInit
+	}
 	if err := db.Delete(m).Error; err != nil {
 		return err
 	}
-	cache.Del(context.Background(), m.GetCacheKey())
+	if cache != nil {
+		cache.Del(context.Background(), m.GetCacheKey())
+	}
 	return nil
 }
 
 func Save(m model.Modeler) error {
+	if db == nil {
+		return ErrDBNotInit
+	}
 	if err := db.Save(m).Error; err != nil {
 		return err
 	}
-	cache.Set(context.Background(), m.GetCacheKey(), m, m.GetCacheDuration())
+	if cache != nil {
+		cache.Set(context.Background(), m.GetCacheKey(), m, m.GetCacheDuration())
+	}
 	return nil
 }
 
 func BatchGet(ms []model.Modeler) error {
 	backToDBModelers := make([]model.Modeler, 0)
-	for _, m := range ms {
-		if err := cache.Get(context.Background(), m.GetCacheKey()).Scan(m); err != nil {
-			backToDBModelers = append(backToDBModelers, m)
+	if cache != nil {
+		for _, m := range ms {
+			if err := cache.Get(context.Background(), m.GetCacheKey()).Scan(m); err != nil {
+				backToDBModelers = append(backToDBModelers, m)
+			}
 		}
 	}
-	if len(backToDBModelers) == 0 {
+
+	if cache != nil && len(backToDBModelers) == 0 {
 		return nil
 	}
 	for _, backToDBModeler := range backToDBModelers {
 		if err := db.Find(backToDBModeler).Error; err != nil {
 			return err
 		}
-		cache.Set(context.Background(), backToDBModeler.GetCacheKey(), backToDBModeler, time.Minute)
+		if cache != nil {
+			cache.Set(context.Background(), backToDBModeler.GetCacheKey(), backToDBModeler, time.Minute)
+		}
 	}
 	return nil
 }
